requirements.txt
selenium==4.8.0
mysql-connector==2.2.9